refactor: move Toolbar component to a dedicated toolbar directory
